feat: implement Badge and Manual Add Minute component for total focus management
